Move Select implementation. #2
