+++ src/select.js
@@ -1,9 +1,9 @@
 import { createElement } from 'react'
 import { string } from 'prop-types'
-import createControl from '../create-control'
+import createControl from './create-control'
 
 export default createControl(({ control, ...props }) =>
-  createElement('select', { ...props, ...control })
+  createElement('select', { ...control, ...props })
 )({
   displayName: 'Select',
   propTypes: {
